% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{construct}
\alias{construct}
\alias{construct1}
\title{Construction of the epidemic process.}
\usage{
construct(Coo, Inftime, infper, NR, Tmax, Indx, size, M1, nrows)

construct1(Coo, Inftime, Remtime, NR, Tmax, Indx, size, M1, nrows)
}
\arguments{
\item{Coo}{Matrix indicating the locations of all premisses (Longitude,latitude).}

\item{infper}{The infectious period.}

\item{NR, Tmax}{The number of hosts removed and the final observation time.}

\item{Indx, size}{Empty vectors to store the position of premisses in the matrix M1 that
store the process.}

\item{M1}{2NX5 Matrix to store the dynamic of the process where colomn:
\enumerate{
\item  Contains the events times
\item  Contains the latitude of the infected premisse.
\item  Contains the longititu of the infected premisse.
\item  Indicates the type of event, particulary we o for infection else the
       index of host removed.
\item  The trajectory of the epidemic ie the size of the infections.
}}

\item{nrows}{Integer to keep the number of rows used in M1 since there might be more
rows than events.}

\item{Remtime, Inftime}{Vectors indicating their corresponding removal and infection
times respectively .}
}
\value{
The dynamic of the epidemic process, the position of each infection, the size
        of the source of infection of each infected premisse and the number of events.
}
\description{
\code{construct} reconvers the dynamic of the epidemic process. It reorders the events
                as they happened'
}

