% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{contains}
\alias{contains}
\title{Check for existence of an element in a vector.}
\usage{
contains(X, z)
}
\arguments{
\item{z}{An element to be checked.}

\item{x}{A vector.}
}
\value{
True if \code{z} is in vector \code{X} and False otherwise.
}
\description{
\code{contains} returns a boolean either true or false.
}
\examples{
distance(10,20,5,10)
}

