% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance}
\alias{distance}
\title{Distance between two locations.}
\usage{
distance(phi1, phi2, lamda1, lamda2)
}
\arguments{
\item{phi1, phi2}{Longititudes of location 1 and 2 respectively.}

\item{lamda1, lamda2}{Latitudes of location 1 and 2 respectively.}
}
\value{
The distance between locations \code{phi1,lamda1} and \code{phi2,lamda2}.
}
\description{
\code{distance} returns the square of the distance between two locations.
}
\examples{
distance(10,20,5,10)
}

