% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{likelihood}
\alias{likelihood}
\title{Likelihood computation using rayleigh distribution as kernel.}
\usage{
likelihood(Coo, Sou_siz, S, dyn, beta, alpha, nrows)
}
\arguments{
\item{Coo, Inftime}{Vectors indicating the locations of all premisses and their corresponding
infection times respectively .}

\item{Sou_siz, S}{Vectors indicating the potential source size, the position of premisses
in the matrix M1 that store the process.}

\item{dyn}{2NX5 Matrix indicating dynamic of the process where colomn:
\enumerate{
\item  Contains the events times
\item  Contains the latitude of the infected premisse.
\item  Contains the longititu of the infected premisse.
\item  Indicates the type of event, particulary we o for infection else the
       index of host removed.
\item  The trajectory of the epidemic ie the size of the infections.
}}

\item{beta, alpha}{The contact and the dispersal parameters.}

\item{nrows}{Integer indicating the number of rows used in dyn since there might be more
rows than events.}
}
\value{
The log-likelihood of the parameters.
}
\description{
\code{likelihood} compute the log-likelihood of the parameters when the kernl is
the rayleigh distribution.
}

