% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprmcmc.R
\docType{package}
\name{pprmcmc}
\alias{pprmcmc}
\alias{pprmcmc-package}
\title{pprmcmc: An implementation of MCMC for system with only infections locations known.}
\description{
The pprmcmc package provides five categories of important functions:
\code{\link{raymcmc}}, \code{\link{expmcmc}}, \code{\link{caumcmc}}, \code{\link{simul}} and
\code{\link{speed}}.
}

