% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speeddist.R
\name{simul}
\alias{simul}
\title{Simulaton of the epidemic process to find the speed of the disease}
\usage{
simul(alpha, beta, tim, gama, Tmax, l)
}
\arguments{
\item{alpha, }{beta, gama indicating the dispersal kernel paremeter,
the contact parameter and the infectious period.}

\item{tim}{A vector of observation times}

\item{Tmax}{Final observation time}

\item{l}{Takes values 1, 2, 3.
\enumerate{
\item indicates the rayleigh kernerl
\item indicates the exponential kernel
\item indicates the cauchy kernel
}}
}
\value{
A list with components:
      \describe{
        \item{epidem}{A five-dimentional matrix giving the  dynamic of the process .
       Each column indicates respectively the times, the x cooedinate, the y-coordinate
       indicator (0 if infection and 1 if removal) and the index of individual
       removed (0 if the event is infection).}
        \item{maxdist}{A vector of maximum distances travelled by the wave on each time in tim.}
       }
}
\description{
Epidemic simulation and maximum distance the wave can travel using gillespsie algorithm.
}
\details{
\code{simul} provide the simulation of the epidemic process and the
      maximum distance the wave can travel at each particular obaservation date.
}
\examples{
# Simulation with rayleigh kernel
alpha=0.00012
beta=0.012
tim=10:325
gama=110
Tmax=325
l=1
res=simul(alpha,beta,tim,gama,Tmax,l)

}

