% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Soursamp}
\alias{Soursamp}
\title{Sample  a source of infection to the current premisse.}
\usage{
Soursamp(rtime, time, ind, siz, ele)
}
\arguments{
\item{rtime, time}{Vector of removals and infections times respectively.}

\item{ind, size, ele}{Integers indicating the index of the premisse in the vector time,
the size of the potential sources and the premisse selected respectively.}
}
\value{
The index of the source of the premisse at the position \code{ele}.
}
\description{
\code{Soursamp} returns the index of the source that infect a premisse.
}

